requirements.txt
cycler==0.10.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.6.2
numpy==1.21.1
Pillow==8.3.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.1
six==1.16.0
tifffile==2021.8.8
